requirements.txt
python-binance
python-telegram-bot
python-dotenv
binance
pandas
pandas-ta
telegram
nest-asyncio
matplotlib
numpy
